Matplotlib library
Matplotlib is a plotting library for the Python programming language and its numerical mathematics extension NumPy. Where it is one of the most popular Python packages used for data visualization. It is a cross-platform library for making 2D plots from data in arrays. Here, in this file some basic idea of Matplotlib library.